removed 'model based' components
By repeating the 'model' component in memory we would end up blowing
cache as for an cecs instance with 3 components & 2 entities we'd have
an array map like this:
	ABCABC

Which would result in us blowing cache when we tried to run systems that
only care about a single component. This issue scales with the number of
components & entities.

Instead we need to revert to extending the component data* memory
blocks.